requirements.txt
pillow
pytesseract
opencv-python
numpy
matplotlib
deskew
torch
transformers
einops
streamlit